+++ source_gen/source_gen/source_gen/lib/src/generator_for_annotation.dart
@@ -4,16 +4,19 @@
 
 import 'dart:async';
 
-import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
 
-import 'constants/reader.dart';
 import 'generator.dart';
 import 'library.dart';
 import 'output_helpers.dart';
 import 'type_checker.dart';
 
-/// A [Generator] that invokes [generateForAnnotatedElement] for every [T].
+/// Extend this type to create a [Generator] that invokes
+/// [generateForAnnotatedElement] for every element in the source file annotated
+/// with [T].
+///
+/// When all annotated elements have been processed, the results will be
+/// combined into a single output with duplicate items collapsed.
 ///
 /// For example, this will allow code generated for all elements which are
 /// annotated with `@Deprecated`:
@@ -36,36 +39,31 @@ abstract class GeneratorForAnnotation<T> extends Generator {
 
   @override
   FutureOr<String> generate(LibraryReader library, BuildStep buildStep) async {
-    var values =
-        await new Stream.fromIterable(library.annotatedWith(typeChecker))
-            .asyncExpand((e) {
-      return normalizeGeneratorOutput(() =>
-          generateForAnnotatedElement(e.element, e.annotation, buildStep));
-    }).fold(new Set<String>(), (Set<String> uniqueValues, value) {
-      value = value?.trim();
+    var values = new Set<String>();
 
-      if (value != null && value.isNotEmpty) {
-        uniqueValues.add(value);
+    for (var annotatedElement in library.annotatedWith(typeChecker)) {
+      var generatedValue =
+          generateForAnnotatedElement(annotatedElement, buildStep);
+      await for (var value in normalizeGeneratorOutput(generatedValue)) {
+        assert(value == null || (value.length == value.trim().length));
+        values.add(value);
       }
-      return uniqueValues;
-    });
+    }
 
     return values.join('\n\n');
   }
 
-  /// Override to return source code to generate for [element].
+  /// Implement to return source code to generate for [annotatedElement].
   ///
   /// This method is invoked based on finding elements annotated with an
-  /// instance of [T]. The [annotation] is provided as a [ConstantReader].
+  /// instance of [T].
   ///
   /// Supported return values include a single [String] or multiple [String]
-  /// instances within an [Iterable] or [Stream].
+  /// instances within an [Iterable] or [Stream]. It is also valid to return a
+  /// [Future] of [String], [Iterable], or [Stream].
   ///
-  /// It is also valid to return a [Future] of any of the above.
-  ///
-  /// Implementations should return `null` when no content is generated.
-  ///
-  /// Empty or whitespace-only [String] instances are also ignored.
+  /// Implementations should return `null` when no content is generated. Empty
+  /// or whitespace-only [String] instances are also ignored.
   generateForAnnotatedElement(
-      Element element, ConstantReader annotation, BuildStep buildStep);
+      AnnotatedElement annotatedElement, BuildStep buildStep);
 }